sentinel_speech/requirements.txt
faster-whisper
torch
numpy
redis
nats-py
silero-vad
onnxruntime
